Refactor thread tracking to avoid using svcGetThreadList
We need to track the threads ourselves now, but that also lets us act
more intelligently. We know when the thread is detached, so we can
return an error in join before calling libctru. We can also implement
the "retval" functionality of `pthread_join`.

We can't just only keep track of the thread ID, because libctru expects
us to use their `Thread` pointer/data structure.